requirements.txt
paho-mqtt
pymysql
python-dotenv